Renamed go module
It turns out that `example.com/broker` isn't really a great name
for the go module that we've made here. To that end, I've changed
it to match this repo.

Additionally, for the sake of sanity, I've extracted the `envOr`
function from the broker package into a separate util package so
it can reasonably be used by anything in the repo, including main.

Finally, this also makes the broker credentials for interacting
with the service itself configurable via the `AUTH_USER` and
`AUTH_PASSWORD` env vars, which default to `admin` and `Test1234`
respectively.